Update services, controllers.

+++ HotelBookingAPI/Dto/RequestDto/RoomDto.cs
@@ -9,7 +9,7 @@ namespace HotelBookingAPI.Dto.RequestDto
         [StringLength(50)]
         public string Name { get; set; }
         [Required]
-        public int HotelId { get; set; }
+        public long HotelId { get; set; }
         [Required]
         public int RoomNumber { get; set; }
         [Required]
